% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d0.R, R/pi0.R
\name{d0}
\alias{d0}
\alias{pi0}
\title{Estimate number or proportion of null hypotheses}
\usage{
d0(p_value, lambda = 0.5)

pi0(p_value, lambda = 0.5)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{lambda}{\link{numeric} vector containing one or
more value(s) for the tuning parameter.}
}
\value{
A \link{numeric} vector the same length as \code{lambda}.
}
\description{
The functions provides Storey's estimator of
the number and proportion of true null hypotheses.
}
\details{
Consider the testing of \eqn{d} null hypotheses with
observed P-values \eqn{p_{1}, \ldots, p_{d}}.\cr\cr

Storey's estimators of the
number and proportion of true null
hypotheses are respectively:
\deqn{\displaystyle{
\hat{d}_0 = \min\left(
 \frac{\sum_{j=1}^{d}{1_{p_j > \lambda} + 1}}
 {(1-\lambda)}, d
\right)
}}

and

\deqn{\displaystyle{
\hat{\pi}_0 = \frac{\hat{d}_0}{d}
}}
}
\references{
Storey, J. D. (2002). A direct approach to false discovery rates.
Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 64(3), 479-498.
}
\author{
Gwenaël G.R. Leday
}
\concept{adaptive}
