% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{decode_special_chars}
\alias{decode_special_chars}
\title{Decode special characters}
\usage{
decode_special_chars(block)
}
\arguments{
\item{block}{Translated description block (list of strings)}
}
\value{
block, with special characters
}
\description{
Decode special characters
}
