% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{encode_special_chars}
\alias{encode_special_chars}
\title{Encode special characters}
\usage{
encode_special_chars(block)
}
\arguments{
\item{block}{Translated description block (list of strings)}
}
\value{
block, without special characters
}
\description{
Encode special characters
}
