% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{firstup}
\alias{firstup}
\title{First letter up}
\usage{
firstup(str_list)
}
\arguments{
\item{str_list}{A string}
}
\value{
str_list, same string, with capitalized first letter
}
\description{
First letter up
}
