% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frr_abh.R
\name{frr_abh}
\alias{frr_abh}
\title{FRR control using adaptive Benjamini-Hochberg (step-up)}
\usage{
frr_abh(p_value, lambda = 0.5, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{lambda}{\link{numeric} vector containing one or
more value(s) for the tuning parameter.}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FRR}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FRR}} using adaptive
Benjamini-Hochberg's step-up procedure and
Storey's plugin estimator
}
\details{
Storey's adaptive version of
Benjamini-Hochberg (BH) procedure
consists in using the decision procedure
described in \link{mtp-package} using:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{\hat{d}_0}{j},
  \ \text{for}\ j=1, \ldots, d.
 },
}\cr
where \eqn{\widehat{d}_0} is Storey's estimator of the
number \eqn{d_0} of true null hypotheses (see \code{\link[=d0]{d0()}}).
\item the adjusted P-values:
\eqn{\qquad\quad
\displaystyle{
  \begin{cases}
  \widetilde{p}_{1} = \min\left( a_j p_{j}, 1\right),\\
  \widetilde{p}_{j} = \min\left( a_j p_{j},
  \widetilde{p}_{j + 1}\right), \ \text{for}\ j = 1, \ldots, d-1
  \end{cases}
 }
 }
\item the adjusted critical values:
\eqn{\qquad
\displaystyle{
  \widetilde{\alpha}_{j} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, d.
} }
}

The BH procedure guarantees
that \eqn{\text{FRR} \leq \alpha} under
independence or positive dependence of P-values.
}
\references{
Storey, J. D. (2002). A direct approach to false
discovery rates. Journal of the Royal Statistical
Society Series B: Statistical Methodology, 64(3), 479-498.\cr
Storey, J. D., Taylor, J. E., & Siegmund, D. (2004).
Strong control, conservative point estimation and
simultaneous conservative consistency of false
discovery rates: a unified approach.
Journal of the Royal Statistical Society Series B:
Statistical Methodology, 66(1), 187-205.
}
\seealso{
Other FRR: 
\code{\link{frr_bh}()}
}
\concept{FRR}
