% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frx_hoch.R
\name{frx_hoch}
\alias{frx_hoch}
\title{FRX control using modified Hochberg (step-up)}
\usage{
frx_hoch(
  p_values,
  d = 0,
  c_value = NULL,
  output = c("p_values", "c_values", "decisions", "factors")
)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{d}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the exceedance level.}

\item{c_value}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the critical value for FRX(d).\cr
Required if \code{output = "c_values"} or \code{output = "decisions"}.}

\item{output}{[ \code{character(1)} ]\cr
The type of value returned by the function:
\itemize{
\item \code{"p_values"} returns the adjusted p-values
\item \code{"c_values"} returns the adjusted critical values
\item \code{"decisions"} returns the rejection decisions
(1 = rejected and 0 = non-rejected)
\item \code{"factors"} returns adjustment factors
}}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{p_values} with:
\itemize{
\item the adjusted p-values if \code{output = "p_values"}
\item the adjusted critical values if \code{output = "c_values"}
\item the rejection decisions (1 = rejected and 0 = non-rejected)
if \code{output = "decisions"}
\item the adjustment factors if \code{output = "factors"}
}
}
\description{
Control FRX(d) using modification of
Hochberg's step-up procedure (Sarkar, 2008;
Guo et al., 2014)
}
\details{
The modified Hochberg procedure
(Guo et al., 2014; Theorem 3.1)
yields:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{d - j + \left\lfloor d j \right\rfloor + 1}
  {\left\lfloor d j \right\rfloor + 1},
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \begin{cases}
  \widetilde{p}_{(m)} = \min\left( a_m p_{(m)}, 1\right),\\
  \widetilde{p}_{(j)} = \min\left( a_j p_{(j)},
  \widetilde{p}_{(j + 1)}\right), \ \text{for}\ j = m - 1, \ldots, 1
  \end{cases}
 }
}
\item the adjusted critical values:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha}_{(j)} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, m.
 }
}
}

Here, \eqn{p_{(1)} \leq \ldots \leq p_{(m)}}
denote the ordered p-values with
\eqn{\widetilde{p}_{(1)}, \ldots,
\widetilde{p}_{(m)}} and
\eqn{\widetilde{\alpha}_{(1)}, \ldots,
\widetilde{\alpha}_{(m)}}
their corresponding adjusted p-values and
critical values. Also,
\eqn{\left\lfloor x \right\rfloor} denotes
the floor function returning the
largest integer that is smaller or equal to \eqn{x}.

The modified Hochberg procedure uses the same
adjustment factors as the modified Holm procedure
(see \code{\link[=frx_holm]{frx_holm()}}) but yields a different
set of rejected hypotheses, as it is a step-up rather
than a step-down procedure.
}
\examples{
# simulate p-values
set.seed(123)
sim <- simulate_p_values(m = 1000, m0 = 900, s1 = 0.1, s2 = 10)

# 1. control FRX(d = 0.1) at 0.05 using adjusted p-values
adj_pv <- frx_hoch(sim$p_value, d = 0.1)
sum(adj_pv <= 0.05)

# 2. control FRX(d = 0.1) at 0.05 using adjusted critical values
adj_cv <- frx_hoch(
  sim$p_value,
  d = 0.1,
  c_value = 0.05,
  output = "c_values"
)
sum(sim$p_value <= adj_cv)

# 3. control FRX(d = 0.1) at 0.05 using decision indicator
ind <- frx_hoch(
  sim$p_value,
  d = 0.1,
  c_value = 0.05,
  output = "decisions"
)
sum(ind)

# 4. adjustment factors to control FRX(d = 0.1) at 0.05
adj_fct <- frx_hoch(sim$p_value, d = 0.1, output = "factors")

}
\references{
Sarkar, S. K. (2008). Generalizing Simes’ test and
Hochberg’s stepup procedure.
The Annals of Statistics, 36(1), 337-363.\cr
Guo, W., He, L., & Sarkar, S. K. (2014). Further results on
controlling the false discovery proportion.
}
\seealso{
Other FRX: 
\code{\link{frx_holm}()}
}
\concept{FRX}
