% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frx_holm.R
\name{frx_holm}
\alias{frx_holm}
\title{FRX control using modified Holm (step-down)}
\usage{
frx_holm(p_value, c = 0, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{c}{\link{numeric} scalar. Exceedance level (between 0 and 1).}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FRX(c)}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FRX(c)}} using modification of
Holm's step-down procedure (Lehmann & Romano, 2005)
}
\details{
The modified Holm procedure
(Lehmann & Romano, 2005; Theorem 3.1)
consists in using the decision procedure
described in \link{mtp-package} using:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{d - j + \left\lfloor c j \right\rfloor + 1}
  {\left\lfloor c j \right\rfloor + 1},
  \ \text{for}\ j=1, \ldots, d.
 }
}
\item the adjusted P-values:
\eqn{\qquad\quad
\displaystyle{
  \begin{cases}
  \widetilde{p}_{1} = \min\left( a_j p_{j}, 1\right),\\
  \widetilde{p}_{j} = \max\left( a_j p_{j},
  \widetilde{p}_{j - 1}\right), \ \text{for}\ j=2, \ldots, d
  \end{cases}
 }
 }
\item the adjusted critical values:
\eqn{\qquad
\displaystyle{
  \widetilde{\alpha}_{j} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, d.
} }
}

Here, \eqn{\left\lfloor x \right\rfloor} denotes
the floor function returning the
largest integer that is smaller or equal to \eqn{x}.

The modified Holm procedure guarantees
that \eqn{\text{FRX(c)} \leq \alpha} without
assumptions on the dependence of P-values.
}
\references{
Lehmann, E. L., & Romano, J. P. (2005). Generalizations of
the familywise error rate. The Annals of Statistics, 33(3), 1138-1154.
}
\seealso{
Other FRX: 
\code{\link{frx_hoch}()}
}
\concept{FRX}
