% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_abon.R
\name{fwer_abon}
\alias{fwer_abon}
\title{FWER control using (single-step) adaptive Bonferroni}
\usage{
fwer_abon(p_value, k = 1, lambda = 0.5, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{k}{\link{numeric} scalar. Exceedance level (positive integer).}

\item{lambda}{\link{numeric} vector containing one or
more value(s) for the tuning parameter.}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FWER(k)}} using generalization of
adaptive Bonferroni's procedure (Wang, 2017)
}
\details{
The generalized adaptive Bonferroni procedure
(Wang, 2017; Definition 2.1)
consists in using the decision procedure
described in \link{mtp-package} with:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \frac{\widehat{d}_0}{k}
 },
}
where \eqn{\widehat{d}_0} is Storey's estimator of the
number \eqn{d_0} of true null hypotheses (see \code{\link[=d0]{d0()}}).
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, d.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}

The generalized adaptive Bonferroni procedure guarantees
that \eqn{\text{FWER(k)} \leq \alpha} without
assumptions on the dependence of P-values.
}
\references{
Storey, J. D. (2002). A direct approach to false discovery rates.
Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 64(3), 479-498.\cr
Wang, L. (2017). Adaptive procedure for generalized
familywise error rate control. Communications in
Statistics-Simulation and Computation, 46(10), 8140-8151.
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER adaptive}
