% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_bon.R
\name{fwer_bon}
\alias{fwer_bon}
\title{FWER control using Bonferroni (single-step)}
\usage{
fwer_bon(
  p_values,
  k = 0,
  c_value = NULL,
  output = c("p_values", "c_values", "decisions", "factors")
)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{k}{[ \code{integer(1)} ]\cr
A non-negative integer smaller than \code{length(p_values)}
specifying the exceedance level.}

\item{c_value}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the critical value for FWER(k).\cr
Required if \code{output = "c_values"} or \code{output = "decisions"}.}

\item{output}{[ \code{character(1)} ]\cr
The type of value returned by the function:
\itemize{
\item \code{"p_values"} returns the adjusted p-values
\item \code{"c_values"} returns the adjusted critical values
\item \code{"decisions"} returns the rejection decisions
(1 = rejected and 0 = non-rejected)
\item \code{"factors"} returns adjustment factors
}}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{p_values} with:
\itemize{
\item the adjusted p-values if \code{output = "p_values"}
\item the adjusted critical values if \code{output = "c_values"}
\item the rejection decisions (1 = rejected and 0 = non-rejected)
if \code{output = "decisions"}
\item the adjustment factors if \code{output = "factors"}
}
}
\description{
Control FWER(k) using generalization of
Bonferroni's single-step procedure.
}
\details{
The generalized Bonferroni procedure
(Lehmann & Romano, 2005; Theorem 2.1)
yields:
\itemize{
\item adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \frac{m}{k + 1}
 }.
}
\item adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}
}
\examples{
# simulate p-values
set.seed(123)
sim <- simulate_p_values(m = 1000, m0 = 900, s1 = 0.1, s2 = 10)

# 1. control FWER(k = 1) at 0.05 using adjusted p-values
adj_pv <- fwer_bon(sim$p_value, k = 1)
sum(adj_pv <= 0.05)

# 2. control FWER(k = 1) at 0.05 using adjusted critical values
adj_cv <- fwer_bon(
  sim$p_value,
  k = 1,
  c_value = 0.05,
  output = "c_values"
)
sum(sim$p_value <= adj_cv)

# 3. control FWER(k = 1) at 0.05 using decision indicator
ind <- fwer_bon(
  sim$p_value,
  k = 1,
  c_value = 0.05,
  output = "decisions"
)
sum(ind)

# 4. adjustment factors to control FWER(k = 1) at 0.05
adj_fct <- fwer_bon(sim$p_value, k = 1, output = "factors")

}
\references{
Bonferroni, C. (1936). Teoria statistica delle classi e
calcolo delle probabilita. Pubblicazioni del R Istituto
Superiore di Scienze Economiche e Commericiali di Firenze, 8, 3-62.\cr
Lehmann, E. L., & Romano, J. P. (2005). Generalizations of
the familywise error rate. The Annals of Statistics, 33(3), 1138-1154.
}
\seealso{
Other FWER: 
\code{\link{fwer_hoch}()},
\code{\link{fwer_holm}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER}
