% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_bon_a.R
\name{fwer_bon_a}
\alias{fwer_bon_a}
\title{FWER control using adaptive Bonferroni (single-step)}
\usage{
fwer_bon_a(
  p_values,
  k = 0,
  lambda = 0.5,
  c_value = NULL,
  output = c("p_values", "c_values", "decisions", "factors")
)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{k}{[ \code{integer(1)} ]\cr
A non-negative integer smaller than \code{length(p_values)}
specifying the exceedance level.}

\item{lambda}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the value for
the tuning parameter of Storey's estimator.}

\item{c_value}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the critical value for FWER(k).\cr
Required if \code{output = "c_values"} or \code{output = "decisions"}.}

\item{output}{[ \code{character(1)} ]\cr
The type of value returned by the function:
\itemize{
\item \code{"p_values"} returns the adjusted p-values
\item \code{"c_values"} returns the adjusted critical values
\item \code{"decisions"} returns the rejection decisions
(1 = rejected and 0 = non-rejected)
\item \code{"factors"} returns adjustment factors
}}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{p_values} with:
\itemize{
\item the adjusted p-values if \code{output = "p_values"}
\item the adjusted critical values if \code{output = "c_values"}
\item the rejection decisions (1 = rejected and 0 = non-rejected)
if \code{output = "decisions"}
\item the adjustment factors if \code{output = "factors"}
}
}
\description{
Control FWER(k) using the generalized
adaptive Bonferroni procedure of Wang (2017).
}
\details{
The generalized adaptive Bonferroni procedure
(Wang, 2017; Definition 2.1)
yields:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \frac{\widehat{m}_0}{k}
 },
}\cr
where \eqn{\widehat{m}_0} is Storey's estimator of the
number \eqn{m_0} of true null hypotheses (see \code{\link[=m0]{m0()}}).
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}
}
\examples{
# simulate p-values
set.seed(123)
sim <- simulate_p_values(m = 1000, m0 = 900, s1 = 0.1, s2 = 10)

# 1. control FWER(k = 1) at 0.05 using adjusted p-values
adj_pv <- fwer_bon_a(sim$p_value, k = 1)
sum(adj_pv <= 0.05)

# 2. control FWER(k = 1) at 0.05 using adjusted critical values
adj_cv <- fwer_bon_a(
  sim$p_value,
  k = 1,
  c_value = 0.05,
  output = "c_values"
)
sum(sim$p_value <= adj_cv)

# 3. control FWER(k = 1) at 0.05 using decision indicator
ind <- fwer_bon_a(
  sim$p_value,
  k = 1,
  c_value = 0.05,
  output = "decisions"
)
sum(ind)

# 4. adjustment factors to control FWER(k = 1) at 0.05
adj_fct <- fwer_bon_a(sim$p_value, k = 1, output = "factors")

}
\references{
Storey, J. D. (2002). A direct approach to false discovery rates.
Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 64(3), 479-498.\cr
Wang, L. (2017). Adaptive procedure for generalized
familywise error rate control. Communications in
Statistics-Simulation and Computation, 46(10), 8140-8151.
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER adaptive}
