% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_bon_w.R
\name{fwer_bon_w}
\alias{fwer_bon_w}
\title{FWER control using weighted Bonferroni (single-step)}
\usage{
fwer_bon_w(
  p_values,
  k = 0,
  weights = NULL,
  c_value = NULL,
  output = c("p_values", "c_values", "decisions", "factors")
)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{k}{[ \code{integer(1)} ]\cr
A non-negative integer smaller than \code{length(p_values)}
specifying the exceedance level.}

\item{weights}{[ \code{numeric()} ]\cr
A numeric vector of weights between 0 and 1
to be allocated to the hypotheses.}

\item{c_value}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the critical value for FWER(k).\cr
Required if \code{output = "c_values"} or \code{output = "decisions"}.}

\item{output}{[ \code{character(1)} ]\cr
The type of value returned by the function:
\itemize{
\item \code{"p_values"} returns the adjusted p-values
\item \code{"c_values"} returns the adjusted critical values
\item \code{"decisions"} returns the rejection decisions
(1 = rejected and 0 = non-rejected)
\item \code{"factors"} returns adjustment factors
}}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{p_values} with:
\itemize{
\item the adjusted p-values if \code{output = "p_values"}
\item the adjusted critical values if \code{output = "c_values"}
\item the rejection decisions (1 = rejected and 0 = non-rejected)
if \code{output = "decisions"}
\item the adjustment factors if \code{output = "factors"}
}
}
\description{
Control FWER(k) using the weighted generalized
Bonferroni procedure of Romano and Wolf (2010).
}
\details{
The weighted generalized Bonferroni procedure
(Romano & Romano, 2010; Theorem 6.1)
yields:
\itemize{
\item adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{1}{w_j (k + 1)}
 }.
}
\item adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a_j\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}
}
\references{
Romano, J. P., & Wolf, M. (2010). Balanced control of
generalized error rates. The Annals of Statistics, 38(1), 598-633.
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER weighted}
