% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_hoch.R
\name{fwer_hoch}
\alias{fwer_hoch}
\title{FWER control using Hochberg (step-up)}
\usage{
fwer_hoch(p_value, k = 1, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{k}{\link{numeric} scalar. Exceedance level (positive integer).}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FWER(k)}} using generalization of
Hochberg's step-up procedure
}
\details{
The generalized Holm procedure
(Sarkar, 2008; Remark 4.2)
consists in using the decision procedure
described in \link{mtp-package} using:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{m - \max\left(j - k, 0\right)}{k},
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted P-values:
\eqn{\qquad\quad
\displaystyle{
  \begin{cases}
  \widetilde{p}_{1} = \min\left( a_j p_{j}, 1\right),\\
  \widetilde{p}_{j} = \min\left( a_j p_{j},
  \widetilde{p}_{j + 1}\right), \ \text{for}\ j = 1, \ldots, m-1
  \end{cases}
 }
 }
\item the adjusted critical values:
\eqn{\qquad
\displaystyle{
  \widetilde{\alpha}_{j} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, m.
} }
}

The generalized Hochberg procedure guarantees
that \eqn{\text{FWER(k)} \leq \alpha} under some
assumptions on the dependence of P-values.

The generalized Hochberg procedure uses the same
adjustment factors as the generalized Holm procedure
(see \code{\link[=fwer_holm]{fwer_holm()}}) but yields a different
set of rejected hypotheses.
}
\references{
Hochberg, Y. (1988). A sharper Bonferroni procedure for
multiple tests of significance. Biometrika, 75(4), 800-802.\cr
Sarkar, S. K. (2008). Generalizing Simes’ test and
Hochberg’s stepup procedure
The Annals of Statistics, 36(1), 337-363.
}
\seealso{
Other FWER: 
\code{\link{fwer_bon}()},
\code{\link{fwer_holm}()}
}
\concept{FWER}
