% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{generate_source_support_blocks}
\alias{generate_source_support_blocks}
\title{Subdivide source-engine tps in blocks}
\usage{
generate_source_support_blocks(
  mtp_directory,
  project_name,
  input_file_name,
  text_identifier_field_name,
  tps
)
}
\arguments{
\item{mtp_directory}{Working directory (string)}

\item{project_name}{Translation project name (string)}

\item{input_file_name}{File name, with path, containing the original descriptions (string)}

\item{text_identifier_field_name}{Field of the input file containing the decription ids (string)}

\item{tps}{Data frame of available translation paths (data frame)}
}
\value{
blocks_file_name, which contains the blocks to be translated (source-support) and the translation plan (plan_file)
}
\description{
Subdivide source-engine tps in blocks
}
