% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{generate_support_target_blocks}
\alias{generate_support_target_blocks}
\title{Subdivide support-engine tps in blocks}
\usage{
generate_support_target_blocks(
  mtp_directory,
  project_name,
  tps,
  target_language
)
}
\arguments{
\item{mtp_directory}{Working directory (string)}

\item{project_name}{Translation project name (string)}

\item{tps}{Data frame of available translation paths (data frame)}

\item{target_language}{Target languange (string)}
}
\value{
blocks_file_name, which contains the blocks to be translated (support-target) and the translation plan (plan_file)
}
\description{
Subdivide support-engine tps in blocks
}
