% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{generate_tps}
\alias{generate_tps}
\title{Define the TPS as a function of the selected languages and engines}
\usage{
generate_tps(
  mtp_directory,
  project_name,
  input_file_name,
  text_identifier_field_name,
  source_languages,
  source_support_engines,
  support_languages,
  support_target_engines,
  target_language
)
}
\arguments{
\item{mtp_directory}{Working directory (string)}

\item{project_name}{Translation project name (string)}

\item{input_file_name}{File name, with path, containing the original descriptions (string)}

\item{text_identifier_field_name}{Name of the field containing the description ids (string)}

\item{source_languages}{A list of string with language ids, e.g. c("en","es","sv")}

\item{source_support_engines}{A list of string with MT tools names, e.g. c("google","deepl")}

\item{support_languages}{A list of string with language ids, e.g. c("en","es","sv")}

\item{support_target_engines}{A list of string with MT tools names, e.g. c("google","deepl")}

\item{target_language}{Target languange (string)}
}
\value{
A data frame containing all the possible TPS
}
\description{
Define the TPS as a function of the selected languages and engines
}
