% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{get_blocks}
\alias{get_blocks}
\title{Subdivide blocks to be submitted to MT}
\usage{
get_blocks(engine, description)
}
\arguments{
\item{engine}{The engine to which submit the descriptions (string)}

\item{description}{A list of strings to be chunked in blocks}
}
\value{
A data frame  in which the descriptions are chunked into blocks
}
\description{
Subdivide blocks to be submitted to MT
}
