% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{getElem}
\alias{getElem}
\title{Equivalent to RSelenium "findElements", but with tryCatch/timeout}
\usage{
getElem(remDr, using, tag)
}
\arguments{
\item{remDr}{The active remote driver (string)}

\item{using}{A string, eg. "xpath", "css selector", see remDr$findElements}

\item{tag}{The webElem to identify (string)}
}
\value{
result, the seeked webElemt (list), empty if not found/timeout
}
\description{
Equivalent to RSelenium "findElements", but with tryCatch/timeout
}
