% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{google_translate}
\alias{google_translate}
\title{Google translate function}
\usage{
google_translate(sl, tl, send, remDr)
}
\arguments{
\item{sl}{Source language (string)}

\item{tl}{Target language (string)}

\item{send}{List of string to be translated}

\item{remDr}{The active remote driver}
}
\value{
A data frame containing the translations of send, plus a checktag "OK" or "KO" related to a successful translation
}
\description{
Google translate function
}
