% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{integrate_translations}
\alias{integrate_translations}
\title{Integrate all the translations for each description, generating the respective translation candidates}
\usage{
integrate_translations(mtp_directory, project_name)
}
\arguments{
\item{mtp_directory}{Working directory (string)}

\item{project_name}{Translation project name (string)}
}
\value{
tcs A data frame with the TCs and the related score
}
\description{
Integrate all the translations for each description, generating the respective translation candidates
}
