% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m0.R, R/pi0.R
\name{m0}
\alias{m0}
\alias{pi0}
\title{Estimate the number or proportion of null hypotheses}
\usage{
m0(p_values, lambda = 0.5)

pi0(p_values, lambda = 0.5)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{lambda}{[ \code{numeric()} ]\cr
A numeric vector containing one or more values
between 0 and 1 for the tuning parameter.}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{lambda}.
}
\description{
The functions provides Storey's estimator of
the number and proportion of true null hypotheses.
}
\details{
Storey's estimators of the
number and proportion of true null
hypotheses are respectively:
\deqn{\displaystyle{
\hat{m}_0 = \min\left(
 \frac{\sum_{j=1}^{m}{1_{p_j > \lambda} + 1}}
 {(1-\lambda)}, m
\right)
}}

and

\deqn{\displaystyle{
\hat{\pi}_0 = \frac{\hat{m}_0}{m}
}}
}
\references{
Storey, J. D. (2002). A direct approach to false discovery rates.
Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 64(3), 479-498.
}
\author{
Gwenaël G.R. Leday
}
\concept{adaptive}
