% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfer_abon.R
\name{pfer_abon}
\alias{pfer_abon}
\title{PFER control using adaptive Bonferroni (single-step)}
\usage{
pfer_abon(p, lambda = 0.5, gamma = NULL, output = "p")
}
\arguments{
\item{p}{\link{numeric} vector. Observed p-values.}

\item{lambda}{\link{numeric} vector containing one or
more value(s) for the tuning parameter.}

\item{gamma}{\link{numeric} scalar. Target \eqn{\text{PFER}}
level (between 1 and m). Overrides \code{output} and
return adjusted critical values.}

\item{output}{\link{character} scalar. Return adjusted p-values
(\code{"p"}) or adjustment factors (\code{"a"}).}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted p-values when \code{output = "p"},
\item adjustment factors when \code{output = "a"},
\item adjusted critical values when \code{gamma} is provided.
}
}
\description{
Control \eqn{\text{PFER}} using adaptive Bonferroni's
single-step procedure
}
\details{
The adaptive Bonferroni procedure
consists in using the decision procedure
described in \link{mtp-package} with:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \widehat{m}_0
 },
}
where \eqn{\widehat{m}_0} is Storey's estimator of the
number \eqn{m_0} of true null hypotheses (see \code{\link[=m0]{m0()}}).
\item the adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\gamma} = \frac{\gamma}{a}.
 }
}
}

The adaptive Bonferroni procedure guarantees
that \eqn{\text{PFER} \leq \gamma} under
independence of p-values.
}
\references{
Guo, W., Lynch, G., & Romano, J. P. (2018).
A new approach for large scale multiple testing with
application to FDR control for graphically
structured hypotheses. arXiv preprint arXiv:1812.00258.
}
\seealso{
Other PFER: 
\code{\link{pfer_bon}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{PFER}
