% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfer_bon.R
\name{pfer_bon}
\alias{pfer_bon}
\title{PFER control using Bonferroni (single-step)}
\usage{
pfer_bon(p, gamma = NULL, output = "p")
}
\arguments{
\item{p}{\link{numeric} vector. Observed p-values.}

\item{gamma}{\link{numeric} scalar. Target \eqn{\text{PFER}}
level (between 1 and m). Overrides \code{output} and
return adjusted critical values.}

\item{output}{\link{character} scalar. Return adjusted p-values
(\code{"p"}) or adjustment factors (\code{"a"}).}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted p-values when \code{output = "p"},
\item adjustment factors when \code{output = "a"},
\item adjusted critical values when \code{gamma} is provided.
}
}
\description{
Control \eqn{\text{PFER}} using Bonferroni's
single-step procedure
}
\details{
The Bonferroni procedure
consists in using the decision procedure
described in \link{mtp-package} with:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = m
 }.
}
\item the adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\gamma} = \frac{\gamma}{a}.
 }
}
}

The Bonferroni procedure guarantees
that \eqn{\text{PFER} \leq \gamma} without
assumptions on the dependence of p-values.
}
\references{
Tukey, J. W. (1953). The problem of multiple comparison.
Unpublished manuscript. In The Collected Works of John W.
Tukey VIII. Multiple Comparisons: 1948–1983 1–300. Chapman
and Hall, New York.\cr
Gordon, A., Glazko, G., Qiu, X., & Yakovlev, A. (2007).
Control of the mean number of false discoveries,
Bonferroni and stability of multiple testing.
The Annals of Applied Statistics, 1(1), 179-190.
}
\seealso{
Other PFER: 
\code{\link{pfer_abon}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{PFER}
