% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfer_bon_a.R
\name{pfer_bon_a}
\alias{pfer_bon_a}
\title{PFER control using adaptive Bonferroni (single-step)}
\usage{
pfer_bon_a(
  p_values,
  lambda = 0.5,
  c_value = NULL,
  output = c("p_values", "c_values", "decisions", "factors")
)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{lambda}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the value for
the tuning parameter of Storey's estimator.}

\item{c_value}{[ \code{numeric(1)} ]\cr
A non-negative integer smaller than \code{length(p_values)}
specifying the critical value for FRR.\cr
Required if \code{output = "c_values"} or \code{output = "decisions"}.}

\item{output}{[ \code{character(1)} ]\cr
The type of value returned by the function:
\itemize{
\item \code{"p_values"} returns the adjusted p-values
\item \code{"c_values"} returns the adjusted critical values
\item \code{"decisions"} returns the rejection decisions
(1 = rejected and 0 = non-rejected)
\item \code{"factors"} returns adjustment factors
}}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{p_values} with:
\itemize{
\item the adjusted p-values if \code{output = "p_values"}
\item the adjusted critical values if \code{output = "c_values"}
\item the rejection decisions (1 = rejected and 0 = non-rejected)
if \code{output = "decisions"}
\item the adjustment factors if \code{output = "factors"}
}
}
\description{
Control PFER using adaptive Bonferroni's
single-step procedure
}
\details{
The Bonferroni procedure
(Meng et al, 2014; Theorem 2.1)
yields:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \widehat{m}_0
 },
}
where \eqn{\widehat{m}_0} is Storey's estimator of the
number \eqn{m_0} of true null hypotheses (see \code{\link[=m0]{m0()}}).
\item the adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\gamma} = \frac{\gamma}{a}.
 }
}
}
}
\references{
Guo, W., Lynch, G., & Romano, J. P. (2018).
A new approach for large scale multiple testing with
application to FDR control for graphically
structured hypotheses. arXiv preprint arXiv:1812.00258.
}
\seealso{
Other PFER: 
\code{\link{pfer_bon}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{PFER}
