% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{rmlastp}
\alias{rmlastp}
\title{Remove last string in between parentheses}
\usage{
rmlastp(str_list)
}
\arguments{
\item{str_list}{A string to trim}
}
\value{
str_list, same string, without the last sentence in between parenthesis (cut SNOMED CT hierarchical tag)
}
\description{
Remove last string in between parentheses
}
