% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{rmsphyp}
\alias{rmsphyp}
\title{Remove consequent spaces and hyphens}
\usage{
rmsphyp(str_list)
}
\arguments{
\item{str_list}{List of strings}
}
\value{
str_list without double spaces or hyphens
}
\description{
Remove consequent spaces and hyphens
}
