% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{sample_engines}
\alias{sample_engines}
\title{Sample plan list to minimize consecutive MT engines queries}
\usage{
sample_engines(plan, engine_field)
}
\arguments{
\item{plan}{A translation plan (data frame)}

\item{engine_field}{A string, either "support_target_engine" or "source_support_engine"}
}
\value{
Same plan, but with shuffled rownames to reduce consecutive MT engines queries
}
\description{
Sample plan list to minimize consecutive MT engines queries
}
