% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_p_values.R
\name{simulate_p_values}
\alias{simulate_p_values}
\title{Simulate p-values}
\usage{
simulate_p_values(m, m0, s1, s2)
}
\arguments{
\item{m}{[ \code{integer(1)} ]\cr
Number of hypotheses.}

\item{m0}{[ \code{integer(1)} ]\cr
Number of null hypotheses.}

\item{s1}{[ \code{numeric(1)} ]\cr
A non-negative number specifying the first shape parameter of
the Beta distribution representing the alternative hypothesis.}

\item{s2}{[ \code{numeric(1)} ]\cr
A non-negative number specifying the second shape parameter of
the Beta distribution representing the alternative hypothesis.}
}
\value{
[ \code{tibble} ] \cr \cr
A tibble/data.frame with components:
\itemize{
\item \code{hypothesis}: [\code{integer}] type of hypothesis
(0 = null and 1 = alternative).
\item \code{p_value}:  [\code{numeric}] generated p-value.
}
}
\description{
Generate p-values from a two-component mixture model with
uniform and Beta distributions.
}
\details{
The generating model is:

\deqn{\displaystyle{
\pi_0 Beta(1, 1) + (1 - \pi_0) Beta(s_1, s_2)
}}
where \eqn{w_0 = m_0/m}.
}
\references{
Allison, D. B., Gadbury, G. L., Heo, M.,
Fernández, J. R., Lee, C. K., Prolla, T. A., &
Weindruch, R. (2002). A mixture model approach for
the analysis of microarray gene expression data.
Computational Statistics & Data Analysis, 39(1), 1-20.
}
