% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{translate}
\alias{translate}
\title{Translate function}
\usage{
translate(engine, sl, tl, send, remDr)
}
\arguments{
\item{engine}{The MT engine (string)}

\item{sl}{Source language (string)}

\item{tl}{Target language (string)}

\item{send}{List of string to be translated}

\item{remDr}{The actuve remote driver}
}
\value{
A data frame containing the translations of send, plus a checktag "OK" or "KO" related to a successful translation, and restored special characters
}
\description{
Translate function
}
