% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTP_load.R
\name{translate_source_support_blocks}
\alias{translate_source_support_blocks}
\title{Translate source-engine tps in blocks}
\usage{
translate_source_support_blocks(
  mtp_directory,
  project_name,
  reftime,
  refsleep,
  timeout_translations
)
}
\arguments{
\item{mtp_directory}{Working directory (string)}

\item{project_name}{Translation project name (string)}

\item{reftime}{Refreshing time, numeric. Default 60 (seconds)}

\item{refsleep}{Sleeping time, numeric. Default 60 (seconds)}

\item{timeout_translations}{Default 20 seconds}
}
\value{
A data frame with appending direct translations
}
\description{
Translate source-engine tps in blocks
}
