% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frr_bh.R
\name{frr_bh}
\alias{frr_bh}
\title{FRR control using Benjamini-Hochberg (step-up)}
\usage{
frr_bh(p_value, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FRR}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FRR}} using Benjamini-Hochberg's
step-up procedure
}
\details{
The Benjamini-Hochberg (BH) procedure
(Sarkar, 2008; Remark 4.2)
consists in using the decision procedure
described in \link{mtp-package} using:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{d}{j},
  \ \text{for}\ j=1, \ldots, d.
 }
}
\item the adjusted P-values:
\eqn{\qquad\quad
\displaystyle{
  \begin{cases}
  \widetilde{p}_{1} = \min\left( a_j p_{j}, 1\right),\\
  \widetilde{p}_{j} = \min\left( a_j p_{j},
  \widetilde{p}_{j + 1}\right), \ \text{for}\ j = 1, \ldots, d-1
  \end{cases}
 }
 }
\item the adjusted critical values:
\eqn{\qquad
\displaystyle{
  \widetilde{\alpha}_{j} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, d.
} }
}

The BH procedure guarantees
that \eqn{\text{FRR} \leq \alpha} under
independence or positive dependence of P-values.
}
\references{
Benjamini, Y., & Hochberg, Y. (1995). Controlling the
false discovery rate: a practical and powerful approach to
multiple testing. Journal of the Royal statistical society:
series B (Methodological), 57(1), 289-300.
}
\seealso{
Other FRR: 
\code{\link{frr_abh}()}
}
\concept{FRR}
