% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frx_hoch.R
\name{frx_hoch}
\alias{frx_hoch}
\title{FRX control using modified Hochberg (step-down)}
\usage{
frx_hoch(p, d = 0, alpha = NULL, output = "p")
}
\arguments{
\item{p}{\link{numeric} vector. Observed P-values.}

\item{d}{\link{numeric} scalar. Exceedance level (between 0 and 1).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FRX(d)}}
level (between 0 and 1). Overrides \code{output} and
return adjusted critical values.}

\item{output}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{output = "p"},
\item adjustment factors when \code{output = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FRX(d)}} using modification of
Hochberg's step-up procedure (Sarkar, 2008;
Guo et al., 2014)
}
\details{
The modified Holm procedure
(Guo et al., 2014; Theorem 3.1)
consists in using the decision procedure
described in \link{mtp-package} using:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{d - j + \left\lfloor d j \right\rfloor + 1}
  {\left\lfloor d j \right\rfloor + 1},
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted P-values:
\eqn{\qquad\quad
\displaystyle{
  \begin{cases}
  \widetilde{p}_{1} = \min\left( a_j p_{j}, 1\right),\\
  \widetilde{p}_{j} = \min\left( a_j p_{j},
  \widetilde{p}_{j + 1}\right), \ \text{for}\ j = 1, \ldots, m-1
  \end{cases}
 }
 }
\item the adjusted critical values:
\eqn{\qquad
\displaystyle{
  \widetilde{\alpha}_{j} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, m.
} }
}

Here, \eqn{\left\lfloor x \right\rfloor} denotes
the floor function returning the
largest integer that is smaller or equal to \eqn{x}.

The modified Hochberg procedure guarantees
that \eqn{\text{FRX(d)} \leq \alpha} under some
assumptions on the dependence of P-values
(Simes inequality).

The modified Hochberg procedure uses the same
adjustment factors as the modified Holm procedure
(see \code{\link[=frx_holm]{frx_holm()}}) but yields a different
set of rejected hypotheses.
}
\references{
Sarkar, S. K. (2008). Generalizing Simes’ test and
Hochberg’s stepup procedure.
The Annals of Statistics, 36(1), 337-363.\cr
Guo, W., He, L., & Sarkar, S. K. (2014). Further results on
controlling the false discovery proportion.
}
\seealso{
Other FRX: 
\code{\link{frx_holm}()}
}
\concept{FRX}
