% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_abon.R
\name{fwer_abon}
\alias{fwer_abon}
\title{FWER control using adaptive Bonferroni (single-step)}
\usage{
fwer_abon(p, k = 0, lambda = 0.5, alpha = NULL, output = "p")
}
\arguments{
\item{p}{\link{numeric} vector. Observed p-values.}

\item{k}{\link{numeric} scalar. Exceedance level (integer).}

\item{lambda}{\link{numeric} vector containing one or
more value(s) for the tuning parameter.}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{output} and
return adjusted critical values.}

\item{output}{\link{character} scalar. Return adjusted p-values
(\code{"p"}) or adjustment factors (\code{"a"}).}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{output = "p"},
\item adjustment factors when \code{output = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FWER(k)}} using generalization of
adaptive Bonferroni's procedure (Wang, 2017)
}
\details{
The generalized adaptive Bonferroni procedure
(Wang, 2017; Definition 2.1)
yields:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \frac{\widehat{m}_0}{k}
 },
}\cr
where \eqn{\widehat{m}_0} is Storey's estimator of the
number \eqn{m_0} of true null hypotheses (see \code{\link[=m0]{m0()}}).
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}
}
\references{
Storey, J. D. (2002). A direct approach to false discovery rates.
Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 64(3), 479-498.\cr
Wang, L. (2017). Adaptive procedure for generalized
familywise error rate control. Communications in
Statistics-Simulation and Computation, 46(10), 8140-8151.
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER adaptive}
