% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_bon.R
\name{fwer_bon}
\alias{fwer_bon}
\title{FWER control using Bonferroni (single-step)}
\usage{
fwer_bon(p_value, k = 1, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{k}{\link{numeric} scalar. Exceedance level (positive integer).}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FWER(k)}} using generalization of
Bonferroni's single-step procedure
}
\details{
The generalized Bonferroni procedure
(Lehmann & Romano, 2005; Theorem 2.1)
consists in using the decision procedure
described in \link{mtp-package} with:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \frac{d}{k}
 }.
}
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, d.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}

The generalized Bonferroni procedure guarantees
that \eqn{\text{FWER(k)} \leq \alpha} without
assumptions on the dependence of P-values.
}
\references{
Bonferroni, C. (1936). Teoria statistica delle classi e
calcolo delle probabilita. Pubblicazioni del R Istituto
Superiore di Scienze Economiche e Commericiali di Firenze, 8, 3-62.\cr
Lehmann, E. L., & Romano, J. P. (2005). Generalizations of
the familywise error rate. The Annals of Statistics, 33(3), 1138-1154.
}
\seealso{
Other FWER: 
\code{\link{fwer_hoch}()},
\code{\link{fwer_holm}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER}
