% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_bon.R
\name{fwer_bon}
\alias{fwer_bon}
\title{FWER control using Bonferroni (single-step)}
\usage{
fwer_bon(p, k = 0, alpha = NULL, output = "p")
}
\arguments{
\item{p}{\link{numeric} vector. Observed P-values.}

\item{k}{\link{numeric} scalar. Exceedance level (positive integer).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{output} and
return adjusted critical values.}

\item{output}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{output = "p"},
\item adjustment factors when \code{output = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FWER(k)}} using generalization of
Bonferroni's single-step procedure
}
\details{
The generalized Bonferroni procedure
(Lehmann & Romano, 2005; Theorem 2.1)
yields:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \frac{m}{k}
 }.
}
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}
}
\references{
Bonferroni, C. (1936). Teoria statistica delle classi e
calcolo delle probabilita. Pubblicazioni del R Istituto
Superiore di Scienze Economiche e Commericiali di Firenze, 8, 3-62.\cr
Lehmann, E. L., & Romano, J. P. (2005). Generalizations of
the familywise error rate. The Annals of Statistics, 33(3), 1138-1154.
}
\seealso{
Other FWER: 
\code{\link{fwer_hoch}()},
\code{\link{fwer_holm}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{FWER}
