% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_holm.R
\name{fwer_holm}
\alias{fwer_holm}
\title{FWER control using Holm (step-down)}
\usage{
fwer_holm(
  p_values,
  k = 0,
  c_value = NULL,
  output = c("p_values", "c_values", "decisions", "factors")
)
}
\arguments{
\item{p_values}{[ \code{numeric()} ]\cr
A numeric vector of observed p-values.}

\item{k}{[ \code{integer(1)} ]\cr
A non-negative integer smaller than \code{length(p_values)}
specifying the exceedance level.}

\item{c_value}{[ \code{numeric(1)} ]\cr
A number between 0 and 1 specifying the critical value for FWER(k).\cr
Required if \code{output = "c_values"} or \code{output = "decisions"}.}

\item{output}{[ \code{character(1)} ]\cr
The type of value returned by the function:
\itemize{
\item \code{"p_values"} returns the adjusted p-values
\item \code{"c_values"} returns the adjusted critical values
\item \code{"decisions"} returns the rejection decisions
(1 = rejected and 0 = non-rejected)
\item \code{"factors"} returns adjustment factors
}}
}
\value{
[ \code{numeric()} ] \cr \cr
A numeric vector the same length as \code{p_values} with:
\itemize{
\item the adjusted p-values if \code{output = "p_values"}
\item the adjusted critical values if \code{output = "c_values"}
\item the rejection decisions (1 = rejected and 0 = non-rejected)
if \code{output = "decisions"}
\item the adjustment factors if \code{output = "factors"}
}
}
\description{
Control FWER(k) using generalization of
Holm's step-down procedure (also called the
Bonferroni-Holm or the sequentially rejective
Bonferroni procedure).
}
\details{
The generalized Holm procedure
(Lehmann & Romano, 2005; Theorem 2.2)
yields:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{m - \max\left(j - k - 1, 0\right)}{k + 1},
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted p-values:
\eqn{\qquad\quad
 \displaystyle{
  \begin{cases}
  \widetilde{p}_{(1)} = \min\left( a_1 p_{(1)}, 1\right),\\
  \widetilde{p}_{(j)} = \max\left( a_j p_{(j)},
  \widetilde{p}_{(j - 1)}\right), \ \text{for}\ j = 2, \ldots, m
  \end{cases}
 }
}
\item the adjusted critical values:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha}_{(j)} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, m.
 }
}
}

Here, \eqn{p_{(1)} \leq \ldots \leq p_{(m)}}
denote the ordered p-values with
\eqn{\widetilde{p}_{(1)}, \ldots,
\widetilde{p}_{(m)}} and
\eqn{\widetilde{\alpha}_{(1)}, \ldots,
\widetilde{\alpha}_{(m)}}
their corresponding adjusted p-values and
critical values.
}
\examples{
# simulate p-values
set.seed(123)
sim <- simulate_p_values(m = 1000, m0 = 900, s1 = 0.1, s2 = 10)

# 1. control FWER(k = 1) at 0.05 using adjusted p-values
adj_pv <- fwer_holm(sim$p_value, k = 1)
sum(adj_pv <= 0.05)

# 2. control FWER(k = 1) at 0.05 using adjusted critical values
adj_cv <- fwer_holm(
  sim$p_value,
  k = 1,
  c_value = 0.05,
  output = "c_values"
)
sum(sim$p_value <= adj_cv)

# 3. control FWER(k = 1) at 0.05 using decision indicator
ind <- fwer_holm(
  sim$p_value,
  k = 1,
  c_value = 0.05,
  output = "decisions"
)
sum(ind)

# 4. adjustment factors to control FWER(k = 1) at 0.05
adj_fct <- fwer_holm(sim$p_value, k = 1, output = "factors")

}
\references{
Holm, S. (1979). A simple sequentially rejective
multiple test procedure. Scandinavian journal of
statistics, 65-70.\cr
Lehmann, E. L., & Romano, J. P. (2005).
Generalizations of the familywise error rate.
The Annals of Statistics, 33(3), 1138-1154.
}
\seealso{
Other FWER: 
\code{\link{fwer_bon}()},
\code{\link{fwer_hoch}()}
}
\concept{FWER}
