% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer_holm.R
\name{fwer_holm}
\alias{fwer_holm}
\title{FWER control using Holm (step-down)}
\usage{
fwer_holm(p, k = 0, alpha = NULL, output = "p")
}
\arguments{
\item{p}{\link{numeric} vector. Observed p-values.}

\item{k}{\link{numeric} scalar. Exceedance level (integer).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{output} and
return adjusted critical values.}

\item{output}{\link{character} scalar. Return adjusted p-values
(\code{"p"}) or adjustment factors (\code{"a"}).}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted p-values when \code{output = "p"},
\item adjustment factors when \code{output = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{FWER(k)}} using generalization of
Holm's step-down procedure
}
\details{
The generalized Holm procedure
(Lehmann & Romano, 2005; Theorem 2.2)
yields:
\itemize{
\item the adjustment factors:
\eqn{\qquad\quad
 \displaystyle{
  a_j = \frac{m - \max\left(j - k - 1, 0\right)}{k + 1},
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted p-values:
\eqn{\qquad\quad
\displaystyle{
  \begin{cases}
  \widetilde{p}_{(1)} = \min\left( a_1 p_{(1)}, 1\right),\\
  \widetilde{p}_{(j)} = \max\left( a_j p_{(j)},
  \widetilde{p}_{(j - 1)}\right), \ \text{for}\ j=2, \ldots, m
  \end{cases}
 }
 }
\item the adjusted critical values:
\eqn{\qquad
\displaystyle{
  \widetilde{\alpha}_{j} = \frac{\alpha}{a_j},
  \ \text{for}\ j=1, \ldots, m.
} }
}
}
\references{
Holm, S. (1979). A simple sequentially rejective
multiple test procedure. Scandinavian journal of
statistics, 65-70.\cr
Lehmann, E. L., & Romano, J. P. (2005).
Generalizations of the familywise error rate.
The Annals of Statistics, 33(3), 1138-1154.
}
\seealso{
Other FWER: 
\code{\link{fwer_bon}()},
\code{\link{fwer_hoch}()}
}
\concept{FWER}
