% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-package.R
\docType{package}
\name{mtp-package}
\alias{mtp-package}
\alias{mtp}
\title{mtp: multiple hypothesis testing procedures}
\description{
Control Type I errors (false positives) when
conducting multiple hypothesis tests using
a variety of error criteria and adjustment methods.

Retrieve adjusted p-values, adjusted critical values,
adjustment factors, or rejection decisions.
}
\section{Control the (generalized) familywise error rate}{

\itemize{
\item \code{\link[=fwer_bon]{fwer_bon()}}
\item \code{\link[=fwer_holm]{fwer_holm()}}
\item \code{\link[=fwer_hoch]{fwer_hoch()}}
\item \code{\link[=fwer_bon_a]{fwer_bon_a()}}
\item \code{\link[=fwer_bon_w]{fwer_bon_w()}}
}
}

\section{Control the false rejection exceedance}{

\itemize{
\item \code{\link[=frx_holm]{frx_holm()}}
\item \code{\link[=frx_hoch]{frx_hoch()}}
}
}

\section{Control the false rejection rate}{

\itemize{
\item \code{\link[=frr_bh]{frr_bh()}}
\item \code{\link[=frr_bh_a]{frr_bh_a()}}
}
}

\section{Control the per family error rate}{

\itemize{
\item \code{\link[=pfer_bon]{pfer_bon()}}
\item \code{\link[=pfer_bon_a]{pfer_bon_a()}}
}
}

\section{Estimate the number/proportion of true null hypotheses}{

\itemize{
\item \code{\link[=m0]{m0()}}
\item \code{\link[=pi0]{pi0()}}
}
}

\section{Simulate p-values}{

\itemize{
\item \code{\link[=simulate_p_values]{simulate_p_values()}}
}
}

\author{
\strong{Maintainer}: Gwenaël G.R. Leday \email{gwenael.leday@proton.me} (\href{https://orcid.org/0000-0002-3753-4060}{ORCID})

}
\keyword{internal}
