% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfer_abon.R
\name{pfer_abon}
\alias{pfer_abon}
\title{PFER control using adaptive Bonferroni (single-step)}
\usage{
pfer_abon(p_value, lambda = 0.5, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{lambda}{\link{numeric} vector containing one or
more value(s) for the tuning parameter.}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{PFER}} using adaptive Bonferroni's
single-step procedure
}
\details{
The adaptive Bonferroni procedure
consists in using the decision procedure
described in \link{mtp-package} with:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = \widehat{m}_0
 },
}
where \eqn{\widehat{m}_0} is Storey's estimator of the
number \eqn{m_0} of true null hypotheses (see \code{\link[=m0]{m0()}}).
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, m.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}

The adaptive Bonferroni procedure guarantees
that \eqn{\text{PFER} \leq \alpha} under
independence of P-values.
}
\references{
Guo, W., Lynch, G., & Romano, J. P. (2018).
A new approach for large scale multiple testing with
application to FDR control for graphically
structured hypotheses. arXiv preprint arXiv:1812.00258.
}
\seealso{
Other PFER: 
\code{\link{pfer_bon}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{PFER}
