% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfer_bon.R
\name{pfer_bon}
\alias{pfer_bon}
\title{PFER control using Bonferroni (single-step)}
\usage{
pfer_bon(p_value, .return = "p", alpha = NULL)
}
\arguments{
\item{p_value}{\link{numeric} vector. Observed P-values.}

\item{.return}{\link{character} scalar. Return adjusted P-values
(\code{"p"}) or adjustment factors (\code{"a"}).}

\item{alpha}{\link{numeric} scalar. Target \eqn{\text{FWER(k)}}
level (between 0 and 1). Overrides \code{.return} and
return adjusted critical values.}
}
\value{
A \link{numeric} vector of:
\itemize{
\item adjusted P-values when \code{.return = "p"},
\item adjustment factors when \code{.return = "a"},
\item adjusted critical values when \code{alpha} is provided.
}
}
\description{
Control \eqn{\text{PFER}} using Bonferroni's
single-step procedure
}
\details{
The Bonferroni procedure
consists in using the decision procedure
described in \link{mtp-package} with:
\itemize{
\item the adjustment factor:
\eqn{\qquad\quad
 \displaystyle{
  a = d
 }.
}
\item the adjusted P-values:
\eqn{\qquad\quad
 \displaystyle{
  \widetilde{p}_{j} = \min\left( a\times p_{j}, 1\right),
  \ \text{for}\ j=1, \ldots, d.
 }
}
\item the adjusted critical value:
\eqn{\qquad
 \displaystyle{
  \widetilde{\alpha} = \frac{\alpha}{a}.
 }
}
}

The Bonferroni procedure guarantees
that \eqn{\text{PFER} \leq \alpha} without
assumptions on the dependence of P-values.
}
\references{
Tukey, J. W. (1953). The problem of multiple comparison.
Unpublished manuscript. In The Collected Works of John W.
Tukey VIII. Multiple Comparisons: 1948–1983 1–300. Chapman
and Hall, New York.\cr
Gordon, A., Glazko, G., Qiu, X., & Yakovlev, A. (2007).
Control of the mean number of false discoveries,
Bonferroni and stability of multiple testing.
The Annals of Applied Statistics, 1(1), 179-190.
}
\seealso{
Other PFER: 
\code{\link{pfer_abon}()}
}
\author{
Gwenaël G.R. Leday
}
\concept{PFER}
